requirements.txt
phidata
fastapi
uvicorn
pandas
pyodbc
python-dotenv
streamlit